Refactor DlgSetupManager into modular setupmanager package
