+++ qtdialogs/setupmanager/__init__.py
@@ -1,0 +1,14 @@
+################################################################################
+#                                                                              #
+#  Copyright (C) 2025, goatpig                                                 #
+#  Distributed under the MIT license                                           #
+#  See LICENSE-MIT or https://opensource.org/licenses/MIT                      #
+#                                                                              #
+################################################################################
+#
+# Setup Manager package - exports DlgSetupManager dialog
+#
+
+from qtdialogs.setupmanager.DlgSetupManager import DlgSetupManager
+
+__all__ = ['DlgSetupManager']
